Fix race condition in ReadAsync.
ByteArrayListStream asynchronously set the release object for its
AsyncLock after locking it in its constructor. However, if the
continuation doesn't run immediately, AddByteArray()/Complete() may
try to use the release object before it's set. When that happens the
lock never gets unlocked for ReadAsync().

Avoid the race condition by adding a method to create an AsyncLock that
starts out locked and returns the release object.